Move interfaces into typings file
